requirements.txt
streamlit>=1.20.0
requests
faiss-cpu
numpy
tqdm
PyPDF2
python-dotenv
